+++ Source/Tools/FEXLoader/TestHarnessRunner/HostRunner.cpp
@@ -1,9 +1,9 @@
 #include "LinuxSyscalls/SignalDelegator.h"
+#include "LinuxSyscalls/UContext.h"
 #include <FEXCore/Config/Config.h>
 #include <FEXCore/Core/CPUBackend.h>
 #include <FEXCore/Core/Context.h>
 #include <FEXCore/Core/CoreState.h>
-#include <FEXCore/Core/UContext.h>
 #include <FEXCore/Core/X86Enums.h>
 #include <FEXCore/Debug/InternalThreadState.h>
 #include <FEXCore/fextl/list.h>
